+++ wechatpy/pay/__init__.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 from __future__ import absolute_import, unicode_literals
-import copy
+import sys
+import inspect
 
 import requests
 import xmltodict
@@ -12,6 +13,10 @@ from wechatpy.pay.utils import calculate_signature
 from wechatpy.pay.utils import dict_to_xml
 from wechatpy.pay.base import BaseWeChatPayAPI
 from wechatpy.pay import api
+
+
+def _is_api_endpoint(obj):
+    return isinstance(obj, BaseWeChatPayAPI)
 
 
 class WeChatPay(object):
@@ -33,10 +38,20 @@ class WeChatPay(object):
 
     def __new__(cls, *args, **kwargs):
         self = super(WeChatPay, cls).__new__(cls)
-        for name, _api in self.__class__.__dict__.items():
-            if isinstance(_api, BaseWeChatPayAPI):
-                _api = copy.deepcopy(_api)
-                _api._client = self
+        if sys.version_info[:2] == (2, 6):
+            import copy
+            # Python 2.6 inspect.gemembers bug workaround
+            # http://bugs.python.org/issue1785
+            for name, _api in self.__class__.__dict__.items():
+                if isinstance(_api, BaseWeChatPayAPI):
+                    _api = copy.deepcopy(_api)
+                    _api._client = self
+                    setattr(self, name, _api)
+        else:
+            api_endpoints = inspect.getmembers(self, _is_api_endpoint)
+            for name, _api in api_endpoints:
+                api_cls = type(_api)
+                _api = api_cls(self)
                 setattr(self, name, _api)
         return self
 
@@ -88,10 +103,20 @@ class WeChatPay(object):
             url=url,
             **kwargs
         )
-        res.raise_for_status()
+        try:
+            res.raise_for_status()
+        except requests.RequestException as reqe:
+            raise WeChatPayException(
+                return_code=None,
+                client=self,
+                request=reqe.request,
+                response=reqe.response
+            )
+
         return self._handle_result(res)
 
     def _handle_result(self, res):
+        res.encoding = 'utf-8'
         xml = res.text
         try:
             data = xmltodict.parse(xml)['xml']
@@ -111,7 +136,10 @@ class WeChatPay(object):
                 result_code,
                 return_msg,
                 errcode,
-                errmsg
+                errmsg,
+                client=self,
+                request=res.request,
+                response=res
             )
         return data